Relocate Jacobian and JointState tests (#251)
