+++ python/test/state_representation/test_geometry.py
